Major refactoring of app to avoid circular imports and linter issues. Functional in principle but throws ResolverError exceptions from connexion indicating that endpoints are not added. However, Swagger UI works just fine. Probably has to do with imports. Also, curl POST (from Swagger) do not work, but on UI works fine.
